Pipeline works for sample data, need to resolve remaining bug to be RAMP compatible
